+++ TinkoffASDKUI/TinkoffASDKUI/Modules/CommonSheet/Views/CommonSheetView.swift
@@ -123,22 +123,25 @@ final class CommonSheetView: UIView {
     // MARK: CommonSheetView Updating
 
     func update(state: CommonSheetState, animated: Bool = true) {
-        if animated {
-            showOverlay { [self] _ in
-                updateViews(with: state)
-                layoutIfNeeded()
-                delegate?.commonSheetView(self, didUpdateWithState: state)
-                hideOverlay()
-            }
-        } else {
+        let stateUpdates = { [self] in
             updateViews(with: state)
+            layoutIfNeeded()
             delegate?.commonSheetView(self, didUpdateWithState: state)
+        }
+
+        guard animated else { return stateUpdates() }
+
+        showOverlay { [self] _ in
+            stateUpdates()
+            hideOverlay()
         }
     }
 
     // MARK: Initial Configuration
 
     private func setupView() {
+        backgroundColor = ASDKColors.Background.elevation1.color
+
         setupLayout()
         updateViews(with: CommonSheetState(status: .processing))
     }