+++ TinkoffASDKUI/TinkoffASDKUI/Modules/MainForm/Router/MainFormRouter.swift
@@ -27,7 +27,7 @@ final class MainFormRouter: IMainFormRouter {
 
     // MARK: IMainFormRouter
 
-    func openCardPaymentForm(paymentFlow: PaymentFlow, cards: [PaymentCard], output: ICardPaymentPresenterModuleOutput?) {
+    func openCardPayment(paymentFlow: PaymentFlow, cards: [PaymentCard], output: ICardPaymentPresenterModuleOutput?) {
         let cardPaymentViewController = cardPaymentAssembly.build(
             activeCards: cards,
             paymentFlow: paymentFlow,